refactor(News): no news on home page
