+++ src/constants.cpp
@@ -186,4 +186,4 @@ const wxString ATTACHMENTS_FOLDER_DATABASE = "%DATABASE%";
 const wxString ATTACHMENTS_FOLDER_USERPROFILE = "%USERPROFILE%";
 const wxString ATTACHMENTS_FOLDER_APPDATA = "%APPDATA%";
 
-const wxString INIDB_DISPLAY_INTERNET_NEWS = "DISPLAY_INTERNET_NEWS";
+const wxString INIDB_NEWS_LAST_READ_DATE = "NEWS_LAST_READ_DATE";
