+++ src/mmhomepagepanel.h
@@ -68,7 +68,6 @@ private:
     void getData();
     void fillData();
     void OnLinkClicked(wxWebViewEvent& event);
-    const wxString displayWebsiteNews();
     const wxString displayAssets(double& tBalance);
     const wxString displayIncomeVsExpenses();
     const wxString getStatWidget();
@@ -81,15 +80,6 @@ private:
     void getExpensesIncomeStats(std::map<int, std::pair<double, double> > &incomeExpensesStats
         , mmDateRange* date_range) const;
     int countFollowUp_, total_transactions_;
-
-    struct WebsiteNews
-    {
-        wxDateTime Date;
-        wxString Title;
-        wxString Link;
-        wxString Description;
-    };
-    const bool getNewsRSS(std::vector<WebsiteNews>& WebsiteNewsList);
 };
 
 #endif
